// frontend/lib/appstate/features/land/action-types.tsx
export enum ActionType {
  LAND_LOADING = "land/LAND_LOADING",
  CREATE_LAND = "land/CREATE_LAND",
  VERIFY_LAND = "land/VERIFY_LAND",
  SET_SHOW_FILE_UPLOAD = "land/SET_SHOW_FILE_UPLOAD",
  SET_SHOW_lOCATE_LAND = "land/SET_SHOW_LOCATE_LAND",
  SET_FILE_PATH = "land/SET_FILE_PATH",
  SET_LOCATION_ADDRESS = "land/SET_LOCATION_ADDRESS",
  LANDs_LOADING = "land/LANDs_LOADING",
  GET_SELLER_LANDS = "land/GET_SELLER_LANDS",
  GET_ALL_LANDS = "land/GET_ALL_LANDS",
  SET_ACTIVE_LINK = "land/SET_ACTIVE_LINK",
  SET_LAND_AREA = "land/SET_LAND_AREA",
}